MAINT: refactor random variables (#68)
* TST: remove underscores from tests

* REV: remove _random_variables.py

* MAINT: first pass at refactoring random variables

* MAINT: refactor __init__

* BUG: add tfp supported discrete variables

* STY: black
