+++ pymc4/__init__.py
@@ -1,10 +1,10 @@
 """PyMC4."""
 from ._model import *
-from ._random_variables import *
+from .random_variables import *
 
 import tensorflow as _tf
 from ._template_contexts import get_context as _get_context
-from ._random_variables import RandomVariable as _RandomVariable
+from .random_variables import RandomVariable as _RandomVariable
 
 __version__ = "0.0.1"
