+++ pymc4/random_variables/continuous.py
@@ -1,0 +1,92 @@
+"""
+PyMC4 continuous random variables.
+
+Wraps selected tfp.distributions (listed in __all__) as pm.RandomVariables.
+Implements random variables not supported by tfp as distributions.
+"""
+
+# FIXME all RandomVariable classes need docstrings
+# pylint: disable=undefined-all-variable
+import sys
+import tensorflow_probability as tfp
+from tensorflow_probability import distributions as tfd
+
+from .random_variable import RandomVariable
+
+
+class HalfStudentT(RandomVariable):
+    def _base_dist(self, *args, **kwargs):
+        """
+        Half student-T base distribution.
+
+        A HalfStudentT is the absolute value of a StudentT.
+        """
+        return tfd.TransformedDistribution(
+            distribution=tfd.StudentT(*args, **kwargs),
+            bijector=tfp.bijectors.AbsoluteValue(),
+            name="HalfStudentT",
+        )
+
+
+class LogitNormal(RandomVariable):
+    def _base_dist(self, *args, **kwargs):
+        """
+        Logit normal base distribution.
+
+        A LogitNormal is the standard logistic (i.e. sigmoid) of a Normal.
+        """
+        return tfd.TransformedDistribution(
+            distribution=tfd.Normal(*args, **kwargs),
+            bijector=tfp.bijectors.Sigmoid(),
+            name="LogitNormal",
+        )
+
+
+class Weibull(RandomVariable):
+    def _base_dist(self, *args, **kwargs):
+        """
+        Weibull base distribution.
+
+        The inverse of the Weibull bijector applied to a U[0, 1] random
+        variable gives a Weibull-distributed random variable.
+        """
+        return tfd.TransformedDistribution(
+            distribution=tfd.Uniform(low=0.0, high=1.0),
+            bijector=tfp.bijectors.Invert(tfp.bijectors.Weibull(*args, **kwargs)),
+            name="Weibull",
+        )
+
+
+# Random variables that tfp supports as distributions. We wrap these
+# distributions as random variables. Names must match tfp.distributions names
+# exactly.
+tfp_supported = [
+    "Beta",
+    "Cauchy",
+    "Chi2",
+    "Exponential",
+    "Gamma",
+    "Gumbel",
+    "HalfCauchy",
+    "HalfNormal",
+    "InverseGamma",
+    "InverseGaussian",
+    "Kumaraswamy",
+    "Laplace",
+    "LogNormal",
+    "Logistic",
+    "Normal",
+    "Pareto",
+    "StudentT",
+    "Triangular",
+    "Uniform",
+    "VonMises",
+]
+
+# Programmatically wrap tfp.distribtions into pm.RandomVariables
+for dist_name in tfp_supported:
+    setattr(
+        sys.modules[__name__],
+        dist_name,
+        type(dist_name, (RandomVariable,), {"_base_dist": getattr(tfd, dist_name)}),
+    )
