+++ pymc4/tests/test_random_variables.py
@@ -2,64 +2,64 @@
 Tests for PyMC4 random variables
 """
 import pytest
-from .. import _random_variables
+from .. import random_variables
 
 
 def random_variable_args():
     """Provide arguments for each random variable."""
     _random_variable_args = (
-        (_random_variables.Bernoulli, {"probs": 0.5}),
-        (_random_variables.Beta, {"concentration0": 1, "concentration1": 1}),
-        (_random_variables.Binomial, {"total_count": 5.0, "probs": 0.5, "sample": 1}),
-        (_random_variables.Categorical, {"probs": [0.1, 0.5, 0.4]}),
-        (_random_variables.Cauchy, {"loc": 0, "scale": 1}),
-        (_random_variables.Chi2, {"df": 2}),
-        (_random_variables.Constant, {"loc": 3}),
-        (_random_variables.Dirichlet, {"concentration": [1, 2], "sample": [0.5, 0.5]}),
+        (random_variables.Bernoulli, {"probs": 0.5}),
+        (random_variables.Beta, {"concentration0": 1, "concentration1": 1}),
+        (random_variables.Binomial, {"total_count": 5.0, "probs": 0.5, "sample": 1}),
+        (random_variables.Categorical, {"probs": [0.1, 0.5, 0.4]}),
+        (random_variables.Cauchy, {"loc": 0, "scale": 1}),
+        (random_variables.Chi2, {"df": 2}),
+        (random_variables.Constant, {"loc": 3}),
+        (random_variables.Dirichlet, {"concentration": [1, 2], "sample": [0.5, 0.5]}),
         # TODO DiscreteUniform raises a NotImplementedError from tfp.
-        # (_random_variables.DiscreteUniform, {"low": 2, "high": 10, "sample": 5}),
-        (_random_variables.Exponential, {"rate": 1}),
-        (_random_variables.Gamma, {"concentration": 3.0, "rate": 2.0}),
-        (_random_variables.Geometric, {"probs": 0.5, "sample": 10}),
-        (_random_variables.Gumbel, {"loc": 0, "scale": 1}),
-        (_random_variables.HalfCauchy, {"loc": 0, "scale": 1}),
-        (_random_variables.HalfNormal, {"scale": 3.0}),
-        (_random_variables.HalfStudentT, {"loc": 0, "scale": 1, "df": 10}),
-        (_random_variables.InverseGamma, {"concentration": 3, "rate": 2}),
-        (_random_variables.InverseGaussian, {"loc": 1, "concentration": 1}),
-        (_random_variables.Kumaraswamy, {"concentration0": 0.5, "concentration1": 0.5}),
-        (_random_variables.LKJ, {"dimension": 1, "concentration": 1.5, "sample": [[1]]}),
-        (_random_variables.Laplace, {"loc": 0, "scale": 1}),
-        (_random_variables.LogNormal, {"loc": 0, "scale": 1}),
-        (_random_variables.Logistic, {"loc": 0, "scale": 3}),
+        # (random_variables.DiscreteUniform, {"low": 2, "high": 10, "sample": 5}),
+        (random_variables.Exponential, {"rate": 1}),
+        (random_variables.Gamma, {"concentration": 3.0, "rate": 2.0}),
+        (random_variables.Geometric, {"probs": 0.5, "sample": 10}),
+        (random_variables.Gumbel, {"loc": 0, "scale": 1}),
+        (random_variables.HalfCauchy, {"loc": 0, "scale": 1}),
+        (random_variables.HalfNormal, {"scale": 3.0}),
+        (random_variables.HalfStudentT, {"loc": 0, "scale": 1, "df": 10}),
+        (random_variables.InverseGamma, {"concentration": 3, "rate": 2}),
+        (random_variables.InverseGaussian, {"loc": 1, "concentration": 1}),
+        (random_variables.Kumaraswamy, {"concentration0": 0.5, "concentration1": 0.5}),
+        (random_variables.LKJ, {"dimension": 1, "concentration": 1.5, "sample": [[1]]}),
+        (random_variables.Laplace, {"loc": 0, "scale": 1}),
+        (random_variables.LogNormal, {"loc": 0, "scale": 1}),
+        (random_variables.Logistic, {"loc": 0, "scale": 3}),
         (
-            _random_variables.Multinomial,
+            random_variables.Multinomial,
             {"total_count": 4, "probs": [0.2, 0.3, 0.5], "sample": [1, 1, 2]},
         ),
-        (_random_variables.LogitNormal, {"loc": 0, "scale": 1}),
+        (random_variables.LogitNormal, {"loc": 0, "scale": 1}),
         (
-            _random_variables.MultivariateNormalFullCovariance,
+            random_variables.MultivariateNormalFullCovariance,
             {"loc": [1, 2], "covariance_matrix": [[0.36, 0.12], [0.12, 0.36]], "sample": [1, 2]},
         ),
-        (_random_variables.NegativeBinomial, {"total_count": 5, "probs": 0.5, "sample": 5}),
-        (_random_variables.Normal, {"loc": 0, "scale": 1}),
-        (_random_variables.Pareto, {"concentration": 1, "scale": 0.1, "sample": 5}),
-        (_random_variables.Poisson, {"rate": 2}),
-        (_random_variables.StudentT, {"loc": 0, "scale": 1, "df": 10}),
-        (_random_variables.Triangular, {"low": 0.0, "high": 1.0, "peak": 0.5}),
-        (_random_variables.Uniform, {"low": 0, "high": 1}),
-        (_random_variables.VonMises, {"loc": 0, "concentration": 1}),
-        (_random_variables.Weibull, {"scale": 0.1, "concentration": 1.0}),
-        (_random_variables.Wishart, {"df": 3, "scale_tril": [[1]], "sample": [[1]]}),
+        (random_variables.NegativeBinomial, {"total_count": 5, "probs": 0.5, "sample": 5}),
+        (random_variables.Normal, {"loc": 0, "scale": 1}),
+        (random_variables.Pareto, {"concentration": 1, "scale": 0.1, "sample": 5}),
+        (random_variables.Poisson, {"rate": 2}),
+        (random_variables.StudentT, {"loc": 0, "scale": 1, "df": 10}),
+        (random_variables.Triangular, {"low": 0.0, "high": 1.0, "peak": 0.5}),
+        (random_variables.Uniform, {"low": 0, "high": 1}),
+        (random_variables.VonMises, {"loc": 0, "concentration": 1}),
+        (random_variables.Weibull, {"scale": 0.1, "concentration": 1.0}),
+        (random_variables.Wishart, {"df": 3, "scale_tril": [[1]], "sample": [[1]]}),
         (
-            _random_variables.ZeroInflatedBinomial,
+            random_variables.ZeroInflatedBinomial,
             {"mix": 0.2, "total_count": 10, "probs": 0.5, "sample": 0},
         ),
         (
-            _random_variables.ZeroInflatedNegativeBinomial,
+            random_variables.ZeroInflatedNegativeBinomial,
             {"mix": 0.2, "total_count": 10, "probs": 0.5, "sample": 0},
         ),
-        (_random_variables.ZeroInflatedPoisson, {"mix": 0.2, "rate": 2, "sample": 0}),
+        (random_variables.ZeroInflatedPoisson, {"mix": 0.2, "rate": 2, "sample": 0}),
     )
 
     ids = [dist[0].__name__ for dist in _random_variable_args]
